Yao/final/UAE/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
args==0.1.0
attrs==25.3.0
av==15.1.0
blessed==1.21.0
braceexpand==0.1.7
certifi==2025.8.3
charset-normalizer==3.4.3
clint==0.5.1
diffusers==0.35.1
distro==1.9.0
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
ftfy==6.3.1
gpustat==1.1.1
h11==0.16.0
hf-xet==1.1.8
-e git+https://github.com/tgxs002/HPSv2.git@866735ecaae999fa714bd9edfa05aa2672669ee3#egg=hpsv2
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.4
jiter==0.10.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvitop==1.5.3
omegaconf==2.3.0
openai==1.104.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.0.0
pluggy==1.6.0
protobuf==3.20.3
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
termcolor==3.1.0
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.50.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webdataset==1.0.2
wheel==0.45.1
zipp==3.23.0
